+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.3.0 (2023-08-21)
+
+### Features
+
+- add a new query `Ecspanse.Query.list_tags/1` to list a component's tags.
+- add a new query `Ecspanse.Query.list_components/1` to list all components of an entity.
+- add a new command `Ecspanse.Command.clone_entity/1` to clone an entity without its relationships.
+- add a new command `Ecspanse.Command.deep_clone_entity/1` to clone an entity with its descendants.
+
 ## v0.2.1 (2023-08-20)
 
 ### Fixes
@@ -16,7 +25,7 @@
 ### Features
 
 - introducing `Ecspanse.Template.Component` and `Ecspanse.Template.Event` to simplify the creation of related components and events.
-- add a new function `Ecspanse.Query.fetch_component/2` to fetch a system's component by a list of tags.
+- add a new query `Ecspanse.Query.fetch_component/2` to fetch a system's component by a list of tags.
 
 ## v0.1.2 (2023-08-14)